requirements.dev.txt
boto3-stubs[s3,mwaa]
aioboto3